idはint64

+++ domain/model/baseModel.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type BaseModel struct {
-	ID        uint `gorm:"primary_key" sql:"AUTO_INCREMENT"`
+	ID        int64 `gorm:"primary_key" sql:"AUTO_INCREMENT"`
 	CreatedAt time.Time `sql:"not null;type:date"`
 	UpdatedAt time.Time `sql:"not null;type:date"`
 }
